+++ core/moengage_flutter/lib/src/moengage_flutter.dart
@@ -1,4 +1,6 @@
 // ignore_for_file: use_setters_to_change_properties
+import 'dart:convert';
+
 import 'package:moengage_flutter_platform_interface/moengage_flutter_platform_interface.dart';
 
 /// Helper Class to interact with MoEngage SDK
@@ -140,7 +142,8 @@ class MoEngageFlutter {
   /// Tracks a user attribute.
   /// Supported attribute types:
   /// - `String` `int`, `double`, `num`, `bool`
-  /// - `List<String>`, `List<int>`, `List<double>`, `List<num>`
+  /// - `List<String>`, `List<int>`, `List<double>`, `List<num>` , List<bool>
+  /// - Valid JSON Object with [Map] and Valid JSON Array with [List]
   /// [userAttributeValue] - Data of type [dynamic]
   /// [userAttributeName] - Name of User Attribute
   void setUserAttribute(String userAttributeName, dynamic userAttributeValue) {
@@ -148,19 +151,12 @@ class MoEngageFlutter {
       Logger.w('User Attribute Name cannot be empty');
       return;
     }
-    if (userAttributeValue is String ||
-        userAttributeValue is int ||
-        userAttributeValue is double ||
-        userAttributeValue is num ||
-        userAttributeValue is bool ||
-        userAttributeValue is List<String> ||
-        userAttributeValue is List<int> ||
-        userAttributeValue is List<double> ||
-        userAttributeValue is List<num>) {
-      _platform.setUserAttribute(userAttributeName, userAttributeValue, appId);
+    final filteredData = filterSupportedTypes(userAttributeValue);
+    if (filteredData != null) {
+      _platform.setUserAttribute(userAttributeName, filteredData, appId);
     } else {
       Logger.w(
-          'Only String, Numbers, Bool and List of Strings/Numbers(non-optional) values supported as User Attributes, provided name: $userAttributeName, value: $userAttributeValue');
+          'Only String, Numbers, Bool, List and JSON Object values are supported as User Attributes, provided name: $userAttributeName, value: $userAttributeValue');
     }
   }