+++ core/moengage_flutter_android/CHANGELOG.md
@@ -3,7 +3,7 @@
 # Next Release Date
 
 ## Next Version
-
+- Added support for JSONObject values in user attributes
 - Added Support for Enforcing SDK into Specific Environment
 
 # 18-06-2024